v1.0 - Added file list and file upload - Added FullCalendar and its related API - Added some DTO classes for controller - Deleted some controller classes and some annotation classes - Roles in the team tab become visible

+++ src/main/java/id/ac/pnj/kanban/response/ProjectResponse.java
@@ -1,20 +1,20 @@
 package id.ac.pnj.kanban.response;
 
 public class ProjectResponse {
-    private Double percentage;
+    private int percentage;
 
     public ProjectResponse() {
     }
 
-    public ProjectResponse(Double percentage) {
+    public ProjectResponse(int percentage) {
         this.percentage = percentage;
     }
 
-    public Double getPercentage() {
+    public int getPercentage() {
         return percentage;
     }
 
-    public void setPercentage(Double percentage) {
+    public void setPercentage(int percentage) {
         this.percentage = percentage;
     }
 }
